Show videos
Added functionality to quit game and load a video. From an admin page (`/admin`), it is possible to start a video using a button. When this button is clicked, a POST is requested to the server, the server then emits a socket, and the client running the game will start the videos. When the videos ends, the game starts again.